combine examples
