--- build-output-api/isr-api-routes/.vercel/output/config.json
@@ -1,3 +1,0 @@
-{
-  "version": 3
-}
