--- build-output-api/isr-api-routes/.vercel/output/functions/index.func/.vc-config.json
@@ -1,5 +1,0 @@
-{
-  "runtime": "nodejs16.x",
-  "handler": "index.js",
-  "launcherType": "Nodejs"
-}
