--- build-output-api/isr-api-routes/.vercel/output/functions/revalidate.func/.vc-config.json
@@ -1,6 +1,0 @@
-{
-  "runtime": "nodejs16.x",
-  "handler": "index.js",
-  "launcherType": "Nodejs",
-  "shouldAddHelpers": true
-}
