+++ build-output-api/on-demand-isr/.vercel/output/functions/revalidate.func/index.js
@@ -9,9 +9,41 @@ const bypassToken = '87734ad8259d67c3c11747d3e4e112d0'
 // DO NOT allow unauthenticated users to trigger revalidation: a malicious user could trigger boundless revalidations.
 const authToken = 'a13f94f6-a441-47ca-95fc-9f44f3450295';
 
+function revalidate(host, path) {
+  return new Promise(function (resolve, reject) {
+    const options = {
+      hostname: host,
+      port: 443,
+      path,
+      method: 'GET', // MUST be "GET" ("HEAD" and "POST" methods will not work)
+      headers: {
+        'x-prerender-revalidate': bypassToken
+      }
+    }
+    const revalidateRequest = https.request(options, revalidateResponse => {
+      const cacheHeader = revalidateResponse.headers['x-vercel-cache'];
+      if (cacheHeader !== 'REVALIDATED') {
+        console.error(new Error(`Revalidation of ${path} failed.`));
+        reject(new Error(`Revalidation of ${path} failed: "x-vercel-cache" is "${cacheHeader}"`))
+        return
+      }
+
+      resolve()
+    })
+
+    revalidateRequest.on('error', error => {
+      console.error(error);
+      reject(error);
+    })
+
+    revalidateRequest.end();
+  })
+}
+
 module.exports = (req, res) => {
   const params = new URLSearchParams(req.query);
-  // Validate the request is allowed to trigger revalidation
+
+  // Validate that the request is allowed to trigger revalidation
   if (authToken !== params.get('authToken')) {
     res.statusCode = 403
     res.end('Not Authorized')
@@ -19,39 +51,22 @@ module.exports = (req, res) => {
 
   const host = req.headers.host
   const deployedUrl = `https://${host}`
+  const pathToRevalidate = params.get('path');
 
-  const options = {
-    hostname: host,
-    port: 443,
-    path: '/',
-    method: 'GET', // MUST be "GET"; a "HEAD" or "POST" request will not work
-    headers: {
-      'x-prerender-revalidate': bypassToken
-    }
-  }
-  const revalidateRequest = https.request(options, revalidateResponse => {
-    res.setHeader('Content-Type', 'text/html charset=utf-8')
+  res.setHeader('Content-Type', 'text/html charset=utf-8')
 
-    if (revalidateResponse.headers['x-vercel-cache'] !== 'REVALIDATED') {
-      res.statusCode = 500
-      res.end(`
-        <h1>Cache NOT Revalidated!</h1>
-        <p>Failed to revalidate. The \`x-vercel-cache\` header was not set to \`REVALIDATED\`.</p>
-      `)
-    }
-
+  revalidate(host, pathToRevalidate).then(() => {
     res.end(`
-      <h1>Cache Revalidated!</h1>
+      <h1>Cache for "${pathToRevalidate}" Revalidated!</h1>
       <p>Redirecting you back.</p>
       <meta http-equiv="refresh" content="2 url=${deployedUrl}">
     `)
-  })
-
-  revalidateRequest.on('error', error => {
+  }).catch((error) => {
     console.error(error.stack)
     res.statusCode = 500
-    res.end()
-  })
-
-  revalidateRequest.end()
+    res.end(`
+      <h1>Cache for "${pathToRevalidate}" NOT Revalidated!</h1>
+      <p>Failed to revalidate: ${error.message}</p>
+    `)
+  });
 }
